backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
weaviate-client==3.25.3
sentence-transformers==2.2.2
python-dotenv==1.0.0
torch==2.1.1
transformers==4.36.0
numpy==1.24.3
scikit-learn==1.3.2